requirements-base.txt
Django==1.6.1
Flask==0.10.1
Jinja2==2.7.1
Markdown==2.3.1
MarkupSafe==0.18
South==0.8.4
Werkzeug==0.9.4
dj-database-url==0.2.2
dj-static==0.0.5
django-classy-tags==0.4
django-cors-headers==1.1.0
django-extensions==1.2.5
django-filter==0.7
django-mptt==0.6.0
django-notifications-hq==0.6.2
django-password-reset==0.7
django-sekizai==0.7
django-taggit==0.12.2
django-toolbelt==0.0.1
djangorestframework==2.3.14
ecdsa==0.10
ipython==1.1.0
itsdangerous==0.23
pillow==2.2.2
pytz==2013.8
sorl-thumbnail==11.12
static==0.4
wiki==0.0.23
django-floppyforms==1.1.1
PyQRCode==1.2.1
pypng==0.0.18
